# server/docs/swagger.yaml
basePath: /
definitions:
...
  apperr.Error:
    properties:
      code:
        type: string
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  url.CreateURLRequest:
    properties:
      custom_code:
        maxLength: 50
        minLength: 3
        type: string
      expires_at:
        type: string
      original_url:
        type: string
      title:
        type: string
    required:
    - original_url
    type: object
  url.DeleteURLResponse:
    properties:
      message:
        type: string
    type: object
  url.ListURLResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/url.URLResponse'
        type: array
      meta:
        $ref: '#/definitions/url.MetaResponse'
    type: object
  url.MetaResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      search:
        type: string
      sort_by:
        type: string
      sort_direction:
        type: string
      total:
        type: integer
    type: object
...
  url.URLResponse:
    properties:
      click_count:
        type: integer
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      original_url:
        type: string
      short_code:
        type: string
      short_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  url.UpdateURLRequest:
    properties:
      expires_at:
        type: string
      is_active:
        type: boolean
      original_url:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/semmidev/url-shortener
  description: High-Performance Modular Monolith URL Shortener REST API in Go.
  license:
    name: MIT
    url: https://github.com/semmidev/url-shortener/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: URL Shortener API
  version: "1.0"
paths:
...
  /api/v1/urls:
    get:
      parameters:
      - description: Search keyword for title, code, or URL
        in: query
        name: search
        type: string
      - description: Column to sort by (created_at, click_count, title, short_code)
        in: query
        name: sort_by
        type: string
      - description: Sort direction (asc, desc)
        in: query
        name: sort_direction
        type: string
      - description: 'Filter status: 1 (active), 0 (inactive), -1 (all)'
        in: query
        name: active
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/url.ListURLResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.Error'
      security:
      - BearerAuth: []
      summary: List user short URLs with DataTable controls
      tags:
      - URLs
    post:
      consumes:
      - application/json
      parameters:
      - description: Create short URL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/url.CreateURLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/url.URLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apperr.Error'
      security:
      - BearerAuth: []
      summary: Create a short URL
      tags:
      - URLs
  /api/v1/urls/{id}:
    delete:
      parameters:
      - description: Short URL UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/url.DeleteURLResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.Error'
      security:
      - BearerAuth: []
      summary: Delete short URL
      tags:
      - URLs
    get:
      parameters:
      - description: Short URL UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/url.URLResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.Error'
      security:
      - BearerAuth: []
      summary: Get short URL by ID
      tags:
      - URLs
    put:
      consumes:
      - application/json
      parameters:
      - description: Short URL UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/url.UpdateURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/url.URLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.Error'
      security:
      - BearerAuth: []
      summary: Update short URL details
      tags:
      - URLs
...
  /api/v1/urls/{id}/qr:
    get:
      parameters:
      - description: Short URL UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: PNG Image
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.Error'
      security:
      - BearerAuth: []
      summary: Get QR Code image (PNG) for a short URL by ID
      tags:
      - URLs
  /api/v1/urls/{id}/restore:
    post:
      parameters:
      - description: Short URL UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/url.URLResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.Error'
      security:
      - BearerAuth: []
      summary: Restore a soft-deleted short URL
      tags:
      - URLs
securityDefinitions:
  BearerAuth:
    description: Input your Bearer token in the format "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"